requirements.txt
anthropic>=0.18.1
google-generativeai>=0.3.1
mistralai>=0.0.8
openai>=1.12.0